# art_admin_backend/docs/swagger.yaml
basePath: /
definitions:
...
  request.CreateJournalEntryRequest:
    properties:
      content:
        type: string
      images:
        items:
          type: string
        type: array
      is_private:
        type: boolean
      mood_record_id:
        type: integer
      tags:
        items:
          type: string
        type: array
      title:
        type: string
    required:
    - content
    - title
    type: object
...
  request.UpdateJournalEntryRequest:
    properties:
      content:
        type: string
      id:
        type: integer
      images:
        items:
          type: string
        type: array
      is_private:
        type: boolean
      mood_record_id:
        type: integer
      tags:
        items:
          type: string
        type: array
      title:
        type: string
    required:
    - id
    type: object
...
  request.UpdateSentimentScoreRequest:
    properties:
      sentiment_score:
        maximum: 1
        minimum: -1
        type: number
    required:
    - sentiment_score
    type: object
...
  response.JournalEntryDetail:
    properties:
      attachment_url:
        type: string
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      is_public:
        type: boolean
      mood_record_id:
        type: integer
      privacy_setting:
        type: string
      sentiment_score:
        type: number
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  response.JournalEntryItem:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      is_public:
        type: boolean
      mood_record_id:
        type: integer
      sentiment_score:
        type: number
      title:
        type: string
      user_id:
        type: integer
    type: object
  response.JournalStatistics:
    properties:
      analysis_date:
        type: string
      average_word_count:
        type: number
      most_common_mood_type:
        type: string
      total_count:
        type: integer
      total_entries:
        type: integer
      writing_streak:
        type: integer
    type: object
...
  response.PaginatedData:
    properties:
      current:
        description: 当前页码
        type: integer
      records:
        description: 记录列表
      size:
        description: 每页条数
        type: integer
      total:
        description: 总记录数
        type: integer
    type: object
  response.Response:
    properties:
      code:
        description: 状态码
        type: integer
      data:
        description: 数据
      msg:
        description: 消息
        type: string
    type: object
...
host: localhost:48080
info:
  contact: {}
  description: Art Admin Backend API Documentation
  title: Art Admin API
  version: "1.0"
paths:
...
  /api/journal-entry/{id}:
    get:
      consumes:
      - application/json
      description: 获取日记条目详细信息
      parameters:
      - description: 日记条目ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.JournalEntryDetail'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取日记条目详情
      tags:
      - 日记管理
  /api/journal-entry/create:
    post:
      consumes:
      - application/json
      description: 创建新的日记条目
      parameters:
      - description: 日记条目信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CreateJournalEntryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 创建日记条目
      tags:
      - 日记管理
  /api/journal-entry/date-range:
    get:
      consumes:
      - application/json
      description: 根据日期范围获取日记条目列表
      parameters:
      - description: 开始日期(YYYY-MM-DD)
        in: query
        name: startDate
        required: true
        type: string
      - description: 结束日期(YYYY-MM-DD)
        in: query
        name: endDate
        required: true
        type: string
      - default: 1
        description: 当前页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页条数
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PaginatedData'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 根据日期范围获取日记条目
      tags:
      - 日记管理
  /api/journal-entry/delete/{id}:
    delete:
      consumes:
      - application/json
      description: 删除日记条目
      parameters:
      - description: 日记条目ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除日记条目
      tags:
      - 日记管理
  /api/journal-entry/list:
    get:
      consumes:
      - application/json
      description: 分页查询日记条目列表
      parameters:
      - description: 当前页码
        in: query
        minimum: 1
        name: page
        required: true
        type: integer
      - description: 每页条数
        in: query
        maximum: 100
        minimum: 1
        name: pageSize
        required: true
        type: integer
      - description: 关联情绪记录ID
        in: query
        name: moodRecordId
        type: integer
      - description: 是否公开
        in: query
        name: isPublic
        type: boolean
      - description: 开始日期(YYYY-MM-DD)
        in: query
        name: startDate
        type: string
      - description: 结束日期(YYYY-MM-DD)
        in: query
        name: endDate
        type: string
      - description: 是否有附件
        in: query
        name: hasAttachment
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PaginatedData'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取日记条目列表
      tags:
      - 日记管理
  /api/journal-entry/recent:
    get:
      consumes:
      - application/json
      description: 获取最近的日记条目列表
      parameters:
      - default: 10
        description: 返回数量限制
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/response.JournalEntryItem'
                  type: array
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取最近的日记条目
      tags:
      - 日记管理
  /api/journal-entry/search:
    get:
      consumes:
      - application/json
      description: 搜索日记条目
      parameters:
      - description: 搜索关键词
        in: query
        name: keyword
        required: true
        type: string
      - description: 当前页码
        in: query
        minimum: 1
        name: page
        required: true
        type: integer
      - description: 每页条数
        in: query
        maximum: 100
        minimum: 1
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 搜索成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PaginatedData'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 搜索日记条目
      tags:
      - 日记管理
  /api/journal-entry/sentiment/{id}:
    put:
      consumes:
      - application/json
      description: 更新日记条目的情感得分
      parameters:
      - description: 日记条目ID
        in: path
        name: id
        required: true
        type: integer
      - description: 情感得分
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.UpdateSentimentScoreRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 更新情感得分
      tags:
      - 日记管理
  /api/journal-entry/statistics:
    get:
      consumes:
      - application/json
      description: 获取日记统计信息
      parameters:
      - default: 30
        description: 统计天数
        in: query
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.JournalStatistics'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取日记统计信息
      tags:
      - 日记管理
  /api/journal-entry/update:
    put:
      consumes:
      - application/json
      description: 更新日记条目信息
      parameters:
      - description: 日记条目信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.UpdateJournalEntryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 更新日记条目
      tags:
      - 日记管理
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"